refactor: move message reactions related components under message reactions directory
